move copy utlil to separate file
